+++ src/main/java/com/shopping/cart/repository/ProductRepository.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends CrudRepository<ProductEntity, Integer> {
+public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
 
     @Query("select p from ProductEntity p where (:name is null or p.name = :name) " +
                                             "and (:price is null or p.price = :price) " +
